requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
python-bitcoinlib==0.11.0
pyln-client==0.10.1
